Simplify setup to introduce the SFT example
