+++ tox.ini
@@ -1,7 +1,7 @@
 [flake8]
 max-line-length = 99
 statistics = True
-exclude = .venv,venv,build,notebooks,.asv
+exclude = .venv,venv,build,notebooks,.asv,out
 ignore =
     E203,
     W503
